Trekk ut felles overskriftkomponent og slett duplikate filer
