+++ src/frontend/komponenter/Fagsak/Vedtak/VedtakBegrunnelserTabell/VedtakBegrunnelser.tsx
@@ -1,10 +1,5 @@
 import React from 'react';
 
-import styled from 'styled-components';
-
-import { PopoverOrientering } from 'nav-frontend-popover';
-import { Element } from 'nav-frontend-typografi';
-
 import { useApp } from '../../../../context/AppContext';
 import { useBehandling } from '../../../../context/BehandlingContext';
 import { useVedtakBegrunnelser } from '../../../../context/VedtakBegrunnelseContext';
@@ -15,26 +10,12 @@ import { IRestVedtakBegrunnelse } from '../../../../typer/vedtak';
 import { Vedtaksperiode, Vedtaksperiodetype } from '../../../../typer/vedtaksperiode';
 import familieDayjs, { familieDayjsDiff } from '../../../../utils/familieDayjs';
 import { datoformat } from '../../../../utils/formatter';
-import Hjelpetekst44px from './Hjelpetekst44px';
 import VedtakBegrunnelsePanel from './VedtaksBegrunnelsePanel';
+import OverskriftMedHjelpetekst from './OverskriftMedHjelpetekst';
 
 interface IVedtakBegrunnelserTabell {
     åpenBehandling: IBehandling;
 }
-
-const UtbetalingsperioderOverskrift = styled.div`
-    margin: 2.75rem 0;
-    display: flex;
-    align-items: center;
-    text-align: center;
-`;
-
-const StyledHjelpetekst44px = styled(Hjelpetekst44px)`
-    .popover {
-        max-width: 18rem;
-        text-align: left;
-    }
-`;
 
 const VedtakBegrunnelser: React.FC<IVedtakBegrunnelserTabell> = ({ åpenBehandling }) => {
     const { toggles } = useApp();
@@ -74,13 +55,12 @@ const VedtakBegrunnelser: React.FC<IVedtakBegrunnelserTabell> = ({ åpenBehandli
 
     return harVedtaksperioder ? (
         <>
-            <UtbetalingsperioderOverskrift>
-                <Element>Begrunnelser i vedtaksbrev</Element>
-                <StyledHjelpetekst44px
-                    type={PopoverOrientering.Hoyre}
-                    innhold="Her skal du sette begrunnelsestekster for innvilgelse, reduksjon og opphør."
-                />
-            </UtbetalingsperioderOverskrift>
+            <OverskriftMedHjelpetekst
+                overskrift={'Begrunnelser i vedtaksbrev'}
+                hjelpetekst={
+                    'Her skal du sette begrunnelsestekster for innvilgelse, reduksjon og opphør.'
+                }
+            />
             {vedtaksperioderMedBegrunnelseBehov
                 .filter((vedtaksperiode: Vedtaksperiode) => {
                     if (toggles[ToggleNavn.visOpphørsperioder])
